requirements.txt
TikTokLive==6.4.5.post1
python-telegram-bot==22.5
